wip quantized matmul

+++ crates/cubecl-linalg/src/matmul/kernels/matmul/base.rs
@@ -140,7 +140,6 @@ fn matmul_cmma_ref_no_check<R: Runtime, EG: MaybeQuantized, A: Algorithm>(
         lhs_line_size,
         rhs_line_size,
         out_line_size,
-        // TODO consider a quantized field for MatmulProblem
     };
 
     let plane_size = client
@@ -183,7 +182,7 @@ fn matmul_launch_kernel<R: Runtime, EG: MaybeQuantized, A: Algorithm>(
     plane_dim: u32,
 ) -> Result<(), MatmulLaunchError> {
     if EG::QUANTIZED {
-        select_kernel::<SingleMatmulSpec<u8, u16, i32>, R, A>(
+        select_kernel::<SingleMatmulSpec<i8, i8, i32>, R, A>(
             client,
             TensorInputsLaunch::new(
                 lhs.as_tensor_arg(lhs_line_size),
